+++ src/Pages/DetailPage/DetailPageStyle.js
@@ -42,6 +42,7 @@ const Thumbnail = styled.img`
 
 const Content = styled.section`
   margin-top: 2rem;
+  line-height: 2;
 `;
 
 const CommentContainer = styled.div`
@@ -58,46 +59,6 @@ const CommentCount = styled.h4`
   font-weight: 600;
   margin-top: 1.33rem;
   margin-bottom: 1.33rem;
-`;
-
-const CommentTextArea = styled.textarea`
-  height: 70px;
-  resize: none;
-  padding: 1rem 1rem 1.5rem;
-  outline: none;
-  border: 1px solid rgb(233, 236, 239);
-  margin-bottom: 1.5rem;
-  width: 100%;
-  border-radius: 4px;
-  min-height: 3.5rem;
-  font-size: 1rem;
-  color: rgb(33, 37, 41);
-  line-height: 1.75;
-`;
-
-const ButtonWraper = styled.div`
-  display: flex;
-  justify-content: flex-end;
-`;
-
-const Button = styled.button`
-  display: inline-flex;
-  align-items: center;
-  justify-content: center;
-  font-weight: bold;
-  cursor: pointer;
-  outline: none;
-  border: none;
-  background: rgb(18, 184, 134);
-  color: white;
-  border-radius: 4px;
-  padding: 0px 1.25rem;
-  height: 2rem;
-  font-size: 1rem;
-
-  &:hover {
-    background-color: #20c997;
-  }
 `;
 
 const UserContainer = styled.footer`
@@ -153,9 +114,6 @@ export const style = {
   Content,
   CommentContainer,
   CommentCount,
-  CommentTextArea,
-  ButtonWraper,
-  Button,
   CommentList,
   UserContainer,
   UserImg,
